Split tests by classes
